pyproject.toml
[project]
name = "nwhacks2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-cpp-python>=0.3.6",
    "mcp[cli]>=1.2.0",
]
